src/backend/api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2

pydantic==2.12.1
pydantic-settings==2.12.0
python-dotenv==1.0.0

sqlalchemy==2.0.29
psycopg2-binary==2.9.7
alembic==1.12.0

passlib[bcrypt]==1.7.4
bcrypt==4.1.2

python-jose[cryptography]